+++ action.yml
@@ -13,4 +13,4 @@ inputs:
     default: "[]"
 runs:
   using: "node12"
-  main: "index.js"
+  main: "index.mjs"
