+++ package.json
@@ -2,7 +2,7 @@
   "name": "pr-reviews-action",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "index.mjs",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
